% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_peaks.R
\name{adjust_peaks}
\alias{adjust_peaks}
\title{Adjust R peaks to hit local max}
\usage{
adjust_peaks(peaks, signal, window = 10)
}
\arguments{
\item{peaks}{Estimated R peaks (as index)}

\item{signal}{ECG signal vector}

\item{window}{Window (+/-) to look for the local peak}
}
\value{
Vector of peaks, adjusted to hit the local max
}
\description{
Adjust the estimated R peaks from \code{\link[rsleep:detect_rpeaks]{rsleep::detect_rpeaks()}} to hit the local maximum.
}
\examples{
data(polar_h10)
bad_segs <- find_bad_segments(polar_h10$time, polar_h10$ecg)
peaks <- detect_peaks(polar_h10$ecg, adjust=FALSE, omit_segments=bad_segs)
peaks_adj <- adjust_peaks(peaks, polar_h10$ecg)
}
\seealso{
\code{\link[=detect_peaks]{detect_peaks()}}
}
