% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rpeaks.R
\name{adjust_rpeaks}
\alias{adjust_rpeaks}
\title{Adjust R peaks to hit local max}
\usage{
adjust_rpeaks(peaks, signal, window = 10)
}
\arguments{
\item{peaks}{Estimated R peaks (as index)}

\item{signal}{ECG signal vector}

\item{window}{Window (+/-) to look for the local peak}
}
\value{
Vector of peaks, adjusted to hit the local max
}
\description{
Adjust the estimated R peaks from \code{\link[rsleep:detect_rpeaks]{rsleep::detect_rpeaks()}} to hit the local maximum.
}
\examples{
data(h10)
peaks <- detect_rpeaks_sliding(h10$ecg)
peaks_adj <- adjust_rpeaks(peaks, h10$ecg)
}
\seealso{
\code{\link[=detect_rpeaks_sliding]{detect_rpeaks_sliding()}}
}
