% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_peaks.R
\name{detect_peaks}
\alias{detect_peaks}
\title{Detect R peaks in a raw ECG signal, using a sliding window}
\usage{
detect_peaks(
  signal,
  window = 80000,
  pad = 15000,
  sRate = 1000000000/7682304,
  ...,
  return_index = TRUE,
  adjust = TRUE
)
}
\arguments{
\item{signal}{Numerical vector of ECG signal}

\item{window}{Integer indicating the number of values to consider at one time}

\item{pad}{Integer indicating the number of values to consider on each side, surrounding the window}

\item{sRate}{ECG signal rate}

\item{...}{Passed to \code{\link[rsleep:detect_rpeaks]{rsleep::detect_rpeaks()}}}

\item{return_index}{If TRUE, the index for each R peak is returned instead of the timing}

\item{adjust}{If TRUE, adjust the results to hit the nearby local maxima}
}
\description{
Detect R peaks in a raw ECG signal, using \code{\link[rsleep:detect_rpeaks]{rsleep::detect_rpeaks()}} but using a sliding window.
}
\examples{
data(h10)
peaks <- detect_peaks(h10$ecg)
}
\seealso{
\code{\link[=adjust_peaks]{adjust_peaks()}}
}
