% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPVC-package.R
\docType{package}
\name{detectPVC-package}
\alias{detectPVC}
\alias{detectPVC-package}
\title{detectPVC: Detect Premature Venticular Complexes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Detect premature ventricular complexes (PVCs) in data from a Polar H10 chest-strap heart rate sensor.
}
\section{Vignette}{

\itemize{
\item \href{https://kbroman.org/detectPVC/}{detectPVC User Guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/detectPVC}
  \item Report bugs at \url{https://github.com/kbroman/detectPVC/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

}
\keyword{internal}
