% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bad_segments.R
\name{find_bad_segments}
\alias{find_bad_segments}
\title{Find bad segments in ECG signal}
\usage{
find_bad_segments(
  time,
  signal,
  absval_thresh = 2,
  runmean_thresh = 0.7,
  window = 0.2,
  min_gap = 2000,
  tz = Sys.timezone()
)
}
\arguments{
\item{time}{Vector of times (integers as from Polar H10, datetimes, or character strings)}

\item{signal}{Numeric vector of ECG signal}

\item{absval_thresh}{Threshold on absolute value of signal}

\item{runmean_thresh}{Threshold on running mean of absolute value of signal}

\item{window}{Window for running mean (in seconds)}

\item{min_gap}{Minimum gap between bad segments (otherwise merged)}

\item{tz}{Timezone used by \code{\link[=convert_timestamp]{convert_timestamp()}}}
}
\value{
Data frame with two columns: the start and end index for
each identified bad segment.
}
\description{
Find particularly noisy regions in ECG signal, with large values far from 0
}
