% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_time.R
\name{find_closest_time}
\alias{find_closest_time}
\title{Find closest time}
\usage{
find_closest_time(target, times, tz = Sys.timezone())
}
\arguments{
\item{target}{The target time, for which we want to find the closest value}

\item{times}{A vector of times}

\item{tz}{Time zone}
}
\value{
Numeric index of the closest time
}
\description{
Find the closest time in a vector of values to a specified target time
}
\examples{
times <- seq(convert_timestamp("2024-05-01 11:00"),
             convert_timestamp("2024-05-01 14:00"), length=300)
find_closest_time("2024-05-01 12:00", times)
}
