% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rpeaks_sliding.R
\name{find_rpeaks_sliding}
\alias{find_rpeaks_sliding}
\title{Detect R peaks in a raw ECG signal, using a sliding window}
\usage{
find_rpeaks_sliding(
  signal,
  window = 250000,
  sRate = 1000000000/7682304,
  ...,
  return_index = TRUE,
  adjust = TRUE
)
}
\arguments{
\item{signal}{Numerical vector of ECG signal}

\item{window}{Integer indicating the number of values to consider at one time}

\item{sRate}{ECG signal rate}

\item{...}{Passed to \code{\link[rsleep:detect_rpeaks]{rsleep::detect_rpeaks()}}}

\item{return_index}{If TRUE, the index for each R peak is returned instead of the timing}

\item{adjust}{If TRUE, adjust the results to hit the nearby local maxima}
}
\description{
Detect R peaks in a raw ECG signal, using \code{\link[rsleep:detect_rpeaks]{rsleep::detect_rpeaks()}} but using a sliding window.
}
\examples{
data(h10)
peaks <- find_rpeaks_sliding(h10$ecg)
}
