% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rpeaks_stats.R
\name{get_rpeaks_stats}
\alias{get_rpeaks_stats}
\title{Get summary statistics for each R peak}
\usage{
get_rpeaks_stats(peaks, signal, window = 10, qtmax = 60)
}
\arguments{
\item{peaks}{Location of R peaks as numeric indexes (from \code{\link[=detect_rpeaks_sliding]{detect_rpeaks_sliding()}} with \code{return_index=TRUE})}

\item{signal}{Vector of ECG signal}

\item{window}{Tight window around each peak to look for local max and min}

\item{qtmax}{Maximum width of QT interval to consider}
}
\value{
A data.frame with local max and min of ECG around peak,
height of T (local max before the next peak), and RR intervals
before and after this R peak.
}
\description{
Get summary statistics for each R peak from ECG data, to help characterize which are normal and which are PVCs
}
\examples{
data(h10)
peaks <- detect_rpeaks_sliding(h10$ecg)
peakstats <- get_rpeaks_stats(peaks, h10$ecg)
}
