% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_interval.R
\name{get_time_interval}
\alias{get_time_interval}
\title{Get indexes for a interval in time}
\usage{
get_time_interval(
  times,
  start = NULL,
  end = NULL,
  length = NULL,
  tz = Sys.timezone()
)
}
\arguments{
\item{times}{A vector of times, either as integers (Polar H10 times in 1e-9 seconds),
as datetimes, or as character strings of the form "2024-05-02 08:56" or
"2024-05-02 08:56:00"}

\item{start}{Start time of interval}

\item{end}{End time of interval}

\item{length}{Length of interval (in seconds)}

\item{tz}{Time zone}
}
\value{
A vector of numeric indexes, for values in \code{times} that are within the specified interval.
}
\description{
For a vector of times, get the indexes that correspond to a specified interval
}
\details{
Two of \code{start}, \code{end}, and \code{length}.
}
\examples{
data(h10)
time_int1 <- get_time_interval(h10$time, start="2024-04-29 22:33:00", length=4)
time_int2 <- get_time_interval(h10$time, start="2024-04-29 22:33:00", end="2024-04-29 22:33:10")
}
\seealso{
\code{\link[=convert_timestamp]{convert_timestamp()}}
}
