% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_common_date.R
\name{most_common_date}
\alias{most_common_date}
\title{Most common date in a vector of date/times}
\usage{
most_common_date(times)
}
\arguments{
\item{Vector}{of date/times}
}
\value{
The most common data, as a character string in ISO format like 2024-05-26
}
\description{
Return the most common date in a vector of date/times
}
\seealso{
\code{\link[=convert_timestamp]{convert_timestamp()}}
}
