% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvc_stats.R
\name{plot.pvc_stats}
\alias{plot.pvc_stats}
\title{Plot running PVC statistics}
\usage{
\method{plot}{pvc_stats}(x, ...)
}
\arguments{
\item{...}{Passed to \code{\link[broman:grayplot]{broman::grayplot()}}}

\item{pvc_stats}{Data frame of running PVC stats from \code{\link[=running_pvc_stats]{running_pvc_stats()}}}
}
\value{
None.
}
\description{
Plot running PVC statistics from \code{\link[=running_pvc_stats]{running_pvc_stats()}}
}
\examples{
data(h10)
peaks <- detect_peaks(h10$ecg)
peak_stats <- calc_peak_stats(peaks, h10$ecg)
pvc <- (peak_stats$RSdist > 6)

pvc_stats <- running_pvc_stats(h10$time, peaks, pvc, window=30, n_at=4)
plot(pvc_stats)
}
