% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_h10-data.R
\docType{data}
\name{polar_h10}
\alias{polar_h10}
\title{Example Polar H10 data}
\format{
A data frame with four columns: time (an epoch-based
integer time stamp in 1e-9 seconds), ecg (the ECG values), hr
(estimated heart rate in beats per minute), and rr (estimated
R-R interval in milliseconds).

There is about 10 minutes of data at about 130 Hz.
}
\source{
Personal data derived from a Polar H10 monitor and extracted using the
\href{https://www.ecglogger.com}{ECGLogger} app.
}
\usage{
data(polar_h10)
}
\description{
Example data from a Polar H10 chest-strap heart rate sensor
}
\examples{
data(polar_h10)
}
\keyword{datasets}
