% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_states.R
\name{print.summary.states}
\alias{print.summary.states}
\title{Print summary of states object}
\usage{
\method{print}{summary.states}(x, digits = 1, ...)
}
\arguments{
\item{x}{Object of class \code{"summary.states"}, as produced by \code{\link[=summary_states]{summary_states()}}.}

\item{digits}{Number of digits in printing; passed to \code{\link[base:print]{base::print()}}}

\item{...}{Ignored}
}
\value{
Invisibly returns the input, \code{x}.
}
\description{
Print summary of states object
}
\examples{
data(polar_h10)
bad_segs <- find_bad_segments(polar_h10$time, polar_h10$ecg)
peaks <- detect_peaks(polar_h10$time, polar_h10$ecg, omit_segments=bad_segs)
peak_stats <- calc_peak_stats(polar_h10$time, polar_h10$ecg, peaks, omit_segments=bad_segs)
pvc <- (peak_stats$RStime > 50)

st <- plot_states(polar_h10$time, peaks, pvc, omit_segments=bad_segs, xlim=c(50, 60), draw=FALSE)
summary(st)
}
\seealso{
\code{\link[=plot_states]{plot_states()}}, \code{\link[=summary_states]{summary_states()}}
}
