% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_datacount.R
\name{running_datacount}
\alias{running_datacount}
\title{Get running count of amount of data}
\usage{
running_datacount(times, window = 1, at = NULL, tz = Sys.timezone())
}
\arguments{
\item{times}{Vector of times}

\item{window}{Length of sliding window in seconds}

\item{at}{Times at which to calculate the amount of data}

\item{tz}{Time zone, to convert \code{times} using \code{\link[=convert_timestamp]{convert_timestamp()}}}
}
\value{
Vector of same length as \code{times}, with the number of data points in a sliding window
centered at each time point.
}
\description{
Get running sum of number of observed data points in sliding windows
}
\examples{
data(h10)
h10$datetime <- convert_timestamp(h10$time)
n_pts <- running_datacount(h10$time)

xax <- broman::time_axis(h10$datetime)
grayplot(h10$datetime, n_pts, xat=NA, vlines=xax$x,
         xlab="Time (seconds)", ylab="No. data points")
axis(side=1, at=xax$x, labels=xax$labels, mgp=c(2.1, 0.5, 0), tick=FALSE)
}
