% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totlength.R
\name{totlength}
\alias{totlength}
\title{Total length of a set of time intervals}
\usage{
totlength(segments, times = NULL, tz = Sys.timezone())
}
\arguments{
\item{segments}{A data frame with two columns, the start and end of
a set of time intervals, either as integer indexes to a
provided vector of times, or as a set of date/times.}

\item{times}{A vector of times, needed if \code{segments} is a set of
integer indexes}

\item{tz}{Time zone, used to convert \code{times} with \code{convert_timestamp}}
}
\value{
Total length of the set of time segments, in seconds
}
\description{
Total length of a set of time intervals, in seconds
}
\examples{
data(h10)
empty_df <- find_bad_segments(h10$time, h10$ecg)

h10$ecg[3896:3931] <- 2.5
badsegs_i <- find_bad_segments(h10$time, h10$ecg)
totlength(badsegs_i, h10$time)

badsegs_t <- find_bad_segments(h10$time, h10$ecg, return_index=FALSE)
totlength(badsegs_t)
}
