% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_segments.R
\name{zero_segments}
\alias{zero_segments}
\title{Zero out segments of a ECG signal}
\usage{
zero_segments(signal, segments)
}
\arguments{
\item{signal}{Numeric vector of ECG signal}

\item{segments}{Data frame with two columns: the start and end index for
each interval that is to have signal replaced by 0}
}
\value{
The input \code{signal} vector, but with the specified intervals replaced with 0.
}
\description{
Replace ECG signal in a set of intervals with 0
}
\examples{
data(h10)
h10$ecg[3896:3931] <- 2.5  # stick in some bad data
badseg <- find_bad_segments(h10$time, h10$ecg)
h10$ecg_z <- zero_segments(h10$ecg, badseg)
}
