% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_timestamp.R
\name{convert_timestamp}
\alias{convert_timestamp}
\title{Convert date/time stamp for Polar H10}
\usage{
convert_timestamp(timestamp, tz = Sys.timezone())
}
\arguments{
\item{timestamp}{A vector of "epoch"-based integers as returned in the Polar H10 data}

\item{tz}{Timezone}
}
\value{
A data.frame with two columns: a DateTime vector and a fraction seconds remainder
}
\description{
Convert the date/time stamp in the Polar H10 data to two columns: date/time and fractional seconds
}
\examples{
data(h10)
ts <- convert_timestamp(h10[,1])
h10 <- cbind(h10, ts)
}
