% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_peaks.R
\name{detect_peaks}
\alias{detect_peaks}
\title{Detect R peaks in a raw ECG signal, using a sliding window}
\usage{
detect_peaks(
  signal,
  window = 80000,
  pad = window/4,
  sRate = 1000000000/7682304,
  peak_limit = 1.5,
  omit_segments = NULL,
  ...,
  return_index = TRUE,
  adjust = TRUE,
  cores = 1
)
}
\arguments{
\item{signal}{Numerical vector of ECG signal}

\item{window}{Integer indicating the number of values to consider at one time}

\item{pad}{Integer indicating the number of values to consider on each side, surrounding the window}

\item{sRate}{ECG signal rate}

\item{peak_limit}{Limit factor on size of peaks relative to mean convolved signal.
Larger values will ignore low-valued peaks.}

\item{omit_segments}{Segments to be omitted: a data frame with two
columns: the start and end index for each interval}

\item{...}{Passed to a modified version of \code{rsleep::detect_rpeaks()}}

\item{return_index}{If TRUE, the index for each R peak is returned instead of the timing}

\item{adjust}{If TRUE, adjust the results to hit the nearby local maxima}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel:detectCores]{parallel::detectCores()}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel:makeCluster]{parallel::makeCluster()}}.}
}
\description{
Detect R peaks in a raw ECG signal, using a modified version of
\code{rsleep::detect_rpeaks()} but using a sliding window.
}
\examples{
data(h10)
peaks <- detect_peaks(h10$ecg)
}
\seealso{
\code{\link[=adjust_peaks]{adjust_peaks()}}
}
