% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bad_segments.R
\name{find_bad_segments}
\alias{find_bad_segments}
\title{Find bad segments in ECG signal}
\usage{
find_bad_segments(
  time,
  signal,
  absval_thresh = 2,
  runmean_thresh = 0.7,
  missing_thresh = 0.1,
  window = 0.2,
  min_gap = 2000,
  pad = 400,
  return_index = TRUE,
  tz = Sys.timezone(),
  sRate = 1000000000/7682304
)
}
\arguments{
\item{time}{Vector of times (integers as from Polar H10, datetimes, or character strings)}

\item{signal}{Numeric vector of ECG signal}

\item{absval_thresh}{Threshold on absolute value of signal}

\item{runmean_thresh}{Threshold on running mean of absolute value of signal}

\item{missing_thresh}{Maximum amount of missing data within 1 second (as a proportion)}

\item{window}{Window for running mean (in seconds)}

\item{min_gap}{Minimum gap between bad segments (otherwise merged)}

\item{pad}{Pad the bad segments by this many values on each end (in index values)}

\item{return_index}{If TRUE, return indexes to \code{time}; otherwise return actual date/times.}

\item{tz}{Timezone used by \code{\link[=convert_timestamp]{convert_timestamp()}}}

\item{sRate}{Signal rate, as expected number of data points per
second; needed if \code{return_prop=TRUE}.}
}
\value{
Data frame with two columns: the start and end index for
each identified bad segment.
}
\description{
Find particularly noisy regions in ECG signal, with large values
far from 0, plus also regions with excessive missing data.
}
\examples{
data(polar_h10)
bad_segs <- find_bad_segments(polar_h10$time, polar_h10$ecg)
}
