% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvc_stats.R
\name{plot.pvc_stats}
\alias{plot.pvc_stats}
\title{Plot running PVC statistics}
\usage{
\method{plot}{pvc_stats}(
  x,
  ylab_pvc = "Percent PVC",
  ylab_hr = "Heart rate (BPM)",
  ylim_pvc = c(0, max(x$pvc, na.rm = TRUE) * 1.02),
  ylim_hr = c(range(x$hr, na.rm = TRUE)) * c(0.98, 1.02),
  ...
)
}
\arguments{
\item{x}{Data frame of running PVC stats from \code{\link[=running_pvc_stats]{running_pvc_stats()}}}

\item{ylab_pvc}{Y-axis label for the plot of percent PVC}

\item{ylab_hr}{Y-axis label for the plot of heart rate}

\item{ylim_pvc}{Y-axis limits for the plot of percent PVC}

\item{ylim_hr}{Y-axis limits for the plot of heart rate}

\item{...}{Passed to \code{\link[broman:grayplot]{broman::grayplot()}}}
}
\value{
None.
}
\description{
Plot running PVC statistics from \code{\link[=running_pvc_stats]{running_pvc_stats()}}
}
\details{
Makes a plot with two panels, of percent PVC vs time and heart rate vs time.
}
\examples{
data(polar_h10)
bad_segs <- find_bad_segments(polar_h10$time, polar_h10$ecg)
peaks <- detect_peaks(polar_h10$ecg, omit_segments=bad_segs)
peak_stats <- calc_peak_stats(polar_h10$time, polar_h10$ecg, peaks, omit_segments=bad_segs)
pvc <- (peak_stats$RStime > 50)

pvc_stats <- running_pvc_stats(polar_h10$time, peaks, pvc, omit_segments=bad_segs,
                               window=60, n_at=100)
plot(pvc_stats)
}
