% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecg.R
\name{plot_ecg}
\alias{plot_ecg}
\title{Plot ECG signal}
\usage{
plot_ecg(
  times,
  signal,
  vlines.col = "#c0a",
  vlines.minor.col = "#f5c",
  bgcolor = "gray98",
  tz = Sys.timezone(),
  ...
)
}
\arguments{
\item{times}{Vector of times (integers as from Polar H10, datetimes, or character strings)}

\item{signal}{ECG signal. If missing or NULL, we take \code{times} as the signal and
then fill in with times assuming Polar H10 signal rate.}

\item{vlines.col}{Color of vertical grid lines at seconds}

\item{vlines.minor.col}{Color of vertical grid lines at 1/5 seconds}

\item{bgcolor}{Background color}

\item{tz}{Timezone (in converting times)}

\item{...}{Passed to \code{\link[broman:grayplot]{broman::grayplot()}}}

\item{return}{None.}
}
\description{
Plot of ECG signal approximately matching traditional ECG graphs
}
\examples{
data(h10)
plot_ecg(h10$ecg)
plot_ecg(h10$time, h10$ecg)
}
