% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecg_mult.R
\name{plot_ecg_mult}
\alias{plot_ecg_mult}
\title{Plot ECG signal across multiple panels}
\usage{
plot_ecg_mult(
  times,
  signal,
  start = NULL,
  length = 30,
  n_panel = 4,
  peaks = NULL,
  pvc = NULL,
  hilit_segments = NULL,
  col_peak = c("slateblue", "violetred"),
  pch_peak = 16,
  cex_peak = 1,
  bg_peak = c("slateblue", "violetred"),
  col_segments = "#c0d3",
  tz = Sys.timezone(),
  ...
)
}
\arguments{
\item{times}{Vector of times (integers as from Polar H10, datetimes, or character strings)}

\item{signal}{ECG signal. If missing or NULL, we take \code{times} as the signal and
then fill in with times assuming Polar H10 signal rate.}

\item{start}{Start time for beginning of first panel, as integer as from Polar H10,
datetimes, or character strings like \code{"2024-05-03 13:27"} or \code{"2024-05-03 13:27:00"}}

\item{length}{Length of time for each panel in seconds}

\item{n_panel}{Number of panels}

\item{peaks}{Optional vector of peak indices; if provided, dots will be plotted at each
peak location}

\item{pvc}{Optional vector of boolean indicators of whether the peaks are PVC or not.
If provided, should be the same length as \code{peaks} and will be used to color the points
at the peaks.}

\item{hilit_segments}{Optional matrix indicating segments to highlight, as returned from
\code{\link[=find_bad_segments]{find_bad_segments()}}, with rows corresponding to intervals and two columns of
numeric indexes of the start and end of each interval}

\item{col_peak}{Vector of two colors to use to color the dots that indicate non-PVC
and PVC peaks, respectively}

\item{pch_peak}{Vector of one or two point types to plot at peaks.}

\item{cex_peak}{Vector of one or two values to indicate size of points
that indicate non-PVC and PVC peaks, respectively}

\item{bg_peak}{Vector of one or two colors to use for background color for
the dots that indicate non-PVC and PVC peaks, respectively.}

\item{col_segments}{Color to highlight the segments in \code{hilit_segments}}

\item{tz}{Timezone (in converting times)}

\item{...}{Passed \code{\link[=plot_ecg]{plot_ecg()}}}
}
\value{
None.
}
\description{
Plot a longer window of ECG signals across multiple panels arranged in one column.
}
\examples{
data(h10)
plot_ecg_mult(h10$time, h10$ecg, start="2024-04-29 22:32", length=20, n_panel=2)

peaks <- detect_peaks(h10$ecg)
peak_stats <- calc_peak_stats(peaks, h10$ecg)
pvc <- (peak_stats$RSdist > 6)
plot_ecg_mult(h10$time, h10$ecg, start="2024-04-29 22:32", length=20, n_panel=2,
              peaks=peaks, pvc=pvc)
}
