% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecg_mult.R
\name{plot_ecg_mult}
\alias{plot_ecg_mult}
\title{Plot ECG signal across multiple panels}
\usage{
plot_ecg_mult(
  times,
  signal,
  start = NULL,
  length = 30,
  n_panel = 4,
  peaks = NULL,
  pvc = NULL,
  hilit_segments = NULL,
  col_peak = c("slateblue", "violetred"),
  pch_peak = 16,
  cex_peak = 1,
  bg_peak = c("slateblue", "violetred"),
  col_segments = "#cc00dd33",
  tz = Sys.timezone(),
  ...
)
}
\arguments{
\item{times}{Vector of times (integers as from Polar H10, datetimes, or character strings)}

\item{signal}{ECG signal. If missing or NULL, we take \code{times} as the signal and
then fill in with times assuming Polar H10 signal rate.}

\item{start}{Start time for beginning of first panel, as integer as from Polar H10,
datetimes, or character strings like \code{"2024-05-03 13:27"} or \code{"2024-05-03 13:27:00"}}

\item{length}{Length of time for each panel in seconds}

\item{n_panel}{Number of panels.}

\item{peaks}{Optional vector of peak indices; if provided, dots will be plotted at each
peak location}

\item{pvc}{Optional vector of boolean indicators of whether the peaks are PVC or not.
If provided, should be the same length as \code{peaks} and will be used to color the points
at the peaks.}

\item{hilit_segments}{Optional matrix indicating segments to highlight, as returned from
\code{\link[=find_bad_segments]{find_bad_segments()}}, with rows corresponding to intervals and two columns of
numeric indexes of the start and end of each interval}

\item{col_peak}{Vector of two colors to use to color the dots that indicate non-PVC
and PVC peaks, respectively}

\item{pch_peak}{Vector of one or two point types to plot at peaks.}

\item{cex_peak}{Vector of one or two values to indicate size of points
that indicate non-PVC and PVC peaks, respectively}

\item{bg_peak}{Vector of one or two colors to use for background color for
the dots that indicate non-PVC and PVC peaks, respectively.}

\item{col_segments}{Color to highlight the segments in \code{hilit_segments}}

\item{tz}{Timezone (in converting times)}

\item{...}{Passed \code{\link[=plot_ecg]{plot_ecg()}}}
}
\value{
None.
}
\description{
Plot a longer window of ECG signals across multiple panels arranged in one column.
}
\details{
We use \code{par(mfrow=c(n_panel, 1))} before creating the set of panels.
However, in the case \code{n_panel==1}, we don't run \code{par(mfrow)} and so this can be
used in a more general way.
}
\examples{
data(polar_h10)
plot_ecg_mult(polar_h10$time, polar_h10$ecg, start="2024-05-05 09:52", length=20, n_panel=2)

bad_segs <- find_bad_segments(polar_h10$time, polar_h10$ecg)
peaks <- detect_peaks(polar_h10$time, polar_h10$ecg, omit_segments=bad_segs)
peak_stats <- calc_peak_stats(polar_h10$time, polar_h10$ecg, peaks, omit_segments=bad_segs)
pvc <- (peak_stats$RStime > 50)
plot_ecg_mult(polar_h10$time, polar_h10$ecg, start="2024-05-05 09:50:30", length=20, n_panel=2,
              peaks=peaks, pvc=pvc, hilit_segments=bad_segs)
}
