% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multcsv.R
\name{read_multcsv}
\alias{read_multcsv}
\title{Read multiple related CSV files}
\usage{
read_multcsv(
  dir = ".",
  files = NULL,
  dont_modify = FALSE,
  omit_incomplete = TRUE,
  tz = Sys.timezone(),
  cores = 1
)
}
\arguments{
\item{dir}{Subdirectory containing the files.}

\item{files}{Optional character vector of file names. If not provided, we use
\code{list.files(dir)} to grab all CSV files in the directory \code{dir}.}

\item{dont_modify}{If TRUE, just rbind the file contents together. If FALSE,
look for a \code{time} column and reorder the rows by that column, and then
add a converted \code{datetime} column using \code{\link[=convert_timestamp]{convert_timestamp()}}.}

\item{omit_incomplete}{If TRUE, omit any rows where the time stamp looks
to be incomplete (having < 19 characters)}

\item{tz}{Time zone, used if \code{dont_modify=FALSE} and there is a time
column to convert.}

\item{cores}{Number of CPU cores to use, for parallel calculations.
(If \code{0}, use \code{\link[parallel:detectCores]{parallel::detectCores()}}.)
Alternatively, this can be links to a set of cluster sockets, as
produced by \code{\link[parallel:makeCluster]{parallel::makeCluster()}}.}
}
\value{
A data.frame with the contents of all files row-binded together.
}
\description{
Read multiple CSV files all with the same columns and rbind them together
}
\details{
At least one of \code{files} or \code{dir} must be provided.
The files should all have the same set of columns.

Files can be gzipped: If a file has name like 'file.csv.gz' it is
gunzipped to a temporary directory and then read.

If \code{dont_modify=FALSE} and there is a column \code{time}, the rows are
reordered using this column and a \code{datetime} column is added,
converting \code{time} with \code{\link[=convert_timestamp]{convert_timestamp()}}
}
