% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multcsv.R
\name{read_multcsv}
\alias{read_multcsv}
\title{Read multiple related CSV files}
\usage{
read_multcsv(files = NULL, dir = ".")
}
\arguments{
\item{files}{Optional character vector of file names. If not provided, we use
\code{list.files(dir)} to grab all CSV files in the directory \code{dir}.}

\item{dir}{Optional subdirectory containing the files.}
}
\value{
A data.frame with the contents of all files row-binded together.
}
\description{
Read multiple CSV files all with the same columns and rbind them together
}
\details{
At least one of \code{files} or \code{dir} must be provided.
The files should all have the same set of columns.
}
