% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_datacount.R
\name{running_datacount}
\alias{running_datacount}
\title{Get running count of amount of data}
\usage{
running_datacount(
  times,
  window = 1,
  at = NULL,
  n_at = NULL,
  tz = Sys.timezone()
)
}
\arguments{
\item{times}{Vector of times}

\item{window}{Length of sliding window in seconds}

\item{at}{Times at which to calculate the amount of data}

\item{n_at}{Number of times at which to calculate the amount of data}

\item{tz}{Time zone, to convert \code{times} using \code{\link[=convert_timestamp]{convert_timestamp()}}}
}
\value{
Vector of same length as \code{times}, with the number of data points in a sliding window
centered at each time point.
}
\description{
Get running sum of number of observed data points in sliding windows
}
\details{
If \code{n_at} and \code{at} are both missing, we calculate amount of data at \code{times}.
If \code{at} is missing but \code{n_at} is provided, we use \code{n_at} equally-spaced times across
the observed range.
}
\examples{
data(polar_h10)
polar_h10$datetime <- convert_timestamp(polar_h10$time)
n_pts <- running_datacount(polar_h10$datetime)

broman::timeplot(polar_h10$datetime, n_pts,
                 xlab="Time", ylab="No. data points")
}
